ecoflow_backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1
psycopg[binary]==3.2.9
python-dotenv==1.0.1
requests==2.32.3
openpyxl==3.1.5
google-auth==2.33.0
